+++ pom.xml
@@ -15,8 +15,9 @@
         <selenium.vesion>4.10.0</selenium.vesion>
         <driveer.manager>5.4.1</driveer.manager>
         <selenaid.version>6.15.0</selenaid.version>
-        <testing.version>6.14.3</testing.version>
-        <assertj-core.version>3.11.1</assertj-core.version>
+        <testing.version>7.7.0</testing.version>
+        <assertj-core.version>3.24.2</assertj-core.version>
+        <rest-assured.version>5.3.1</rest-assured.version>
     </properties>
     <dependencies>
         <dependency>
@@ -52,8 +53,38 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>RELEASE</version>
+            <version>4.13.2</version>
             <scope>test</scope>
+        </dependency>
+
+        <!-- https://mvnrepository.com/artifact/io.rest-assured/rest-assured -->
+        <dependency>
+            <groupId>io.rest-assured</groupId>
+            <artifactId>rest-assured</artifactId>
+            <version>${rest-assured.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.rest-assured</groupId>
+            <artifactId>rest-assured</artifactId>
+            <version>5.3.1</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>de.plushnikov.lombok-intellij-plugin</groupId>
+            <artifactId>lombok-plugin</artifactId>
+            <version>0.5</version>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.28</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>42.6.0</version>
         </dependency>
     </dependencies>
 </project>
