+++ src/main/java/Api/config/BaseResponseSpecification.java
@@ -1,0 +1,12 @@
+package Api.config;
+
+import io.restassured.builder.ResponseSpecBuilder;
+import io.restassured.specification.ResponseSpecification;
+import org.apache.http.HttpStatus;
+
+public class BaseResponseSpecification {
+    public ResponseSpecification OK(){
+        return new ResponseSpecBuilder()
+                .expectStatusCode(HttpStatus.SC_OK).build();
+    }
+}
