+++ src/main/java/Api/config/Request.java
@@ -1,0 +1,42 @@
+package Api.config;
+
+import io.restassured.response.Response;
+import io.restassured.response.ValidatableResponse;
+import io.restassured.specification.RequestSpecification;
+import io.restassured.specification.ResponseSpecification;
+
+import static io.restassured.RestAssured.given;
+
+public class Request {
+
+    public Response post(RequestSpecification request, String body, String endpoint){
+       return given()
+               .spec(request)
+               .body(body)
+               .post(endpoint);
+    }
+    public ValidatableResponse post(RequestSpecification request, String body, String endpoint, ResponseSpecification response){
+        return given()
+                .spec(request)
+                .body(body)
+                .post(endpoint)
+                .then()
+                .spec(response);
+    }
+
+    public ValidatableResponse get(RequestSpecification request, String endpoint, String body, ResponseSpecification response){
+        return given()
+                .spec(request).body(body)
+                .get(endpoint)
+                .then()
+                .spec(response);
+    }
+
+    public Response put(){
+        return null;
+    }
+
+    public Response delete(){
+        return null;
+    }
+}
