+++ src/main/java/Api/dto/UserBuilder.java
@@ -1,0 +1,21 @@
+package Api.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+@Builder
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserBuilder {
+    Integer id;
+    String username;
+    String firstName;
+    String lastName;
+    String email;
+    String password;
+    String phone;
+    String userStatus;
+
+}
