+++ src/main/java/Api/endpoints/EndpointsPOST.java
@@ -1,0 +1,16 @@
+package Api.endpoints;
+
+public enum EndpointsPOST {
+    CREATE_USER("/user");
+    private final String endpoint;
+
+
+    EndpointsPOST(String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+    public String getValue() {
+        return endpoint;
+    }
+
+}
