+++ src/test/api.properties
@@ -1,0 +1,1 @@
+ url=https://petstore.swagger.io/
